-h TehDoc plan update

+++ migrations/m181130_082239_vks_sessions_tbl.php
@@ -23,34 +23,34 @@ class m181130_082239_vks_sessions_tbl extends Migration
             'vks_work_time_end' => $this->string(25),
             'vks_duration_work' => $this->integer(),
 
-            'vks_type' => $this->integer()->notNull(),
+            'vks_type' => $this->bigInteger()->notNull(),
             'vks_type_text' => $this->string(255),
-            'vks_place' => $this->integer(),
+            'vks_place' => $this->bigInteger(),
             'vks_place_text' => $this->string(255),
 
-            'vks_subscriber_office' => $this->integer(),
+            'vks_subscriber_office' => $this->bigInteger(),
             'vks_subscriber_office_text' => $this->string(255),
             'vks_subscriber_name' => $this->string(255),
 
-            'vks_order' => $this->integer(),
+            'vks_order' => $this->bigInteger(),
             'vks_order_text' => $this->string(255),                               // колонка с текстовым содержанием
 
             'vks_order_number' => $this->string(255),
             'vks_order_date' => $this->date(),
 
-            'vks_equipment' => $this->integer(),
+            'vks_equipment' => $this->bigInteger(),
             'vks_remarks' => $this->boolean()->defaultValue(0),                   // замечания
 
-            'vks_employee' => $this->integer(),                                         // сотрудник спецсвязи, участв. в приеме
+            'vks_employee' => $this->bigInteger(),                                         // сотрудник спецсвязи, участв. в приеме
             'vks_employee_text' => $this->string(255),
 
-            'vks_subscriber_reg_office' => $this->integer(),                            // региональный абонент
+            'vks_subscriber_reg_office' => $this->bigInteger(),                            // региональный абонент
             'vks_subscriber_reg_office_text' => $this->string(),
             'vks_subscriber_reg_name' => $this->string(255),
 
             'vks_employee_receive_msg' => $this->string(255),
             'vks_receive_msg_datetime' => $this->dateTime(),
-            'vks_employee_send_msg' => $this->integer(),
+            'vks_employee_send_msg' => $this->bigInteger(),
             'vks_employee_send_msg_text' => $this->string(255),
 
             'vks_comments' => $this->text(),
