+++ migrations/m181203_211220_vks_employees_tbl.php
@@ -19,7 +19,7 @@ class m181203_211220_vks_employees_tbl extends Migration
       'rgt' => $this->integer()->notNull(),
       'lvl' => $this->smallInteger(5)->notNull(),
       'name' => $this->string(120)->notNull(),
-      'parent_id' => $this->integer(),
+      'parent_id' => $this->bigInteger(),
       'valid' => $this->boolean()->defaultValue(1),
       'del_reason' => $this->string(255)
     ], $tableOptions);
