+++ migrations/m181225_083505_tehdoc_category_tbl.php
@@ -22,7 +22,7 @@ class m181225_083505_tehdoc_category_tbl extends Migration
       'rgt' => $this->integer()->notNull(),
       'lvl' => $this->smallInteger(5)->notNull(),
       'name' => $this->string(120)->notNull(),
-      'parent_id' => $this->integer(),
+      'parent_id' => $this->bigInteger(),
       'valid' => $this->boolean()->defaultValue(1),
       'del_reason' => $this->string(255)
     ], $tableOptions);
